+++ app/src/main/java/com/growthook/aos/domain/usecase/home/GetSeedsUseCase.kt
@@ -1,4 +1,4 @@
-package com.growthook.aos.domain.usecase
+package com.growthook.aos.domain.usecase.home
 
 import com.growthook.aos.domain.repository.SeedRepository
 import javax.inject.Inject
